backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.0.0
requests==2.31.0
PyPDF2==3.0.1
PyMuPDF==1.23.8
gunicorn==21.2.0 